feat: 完善iconfont打包逻辑&node v12下编译测试

+++ style-compiler/tasks/iconfont.js
@@ -10,29 +10,16 @@ function getSvgs(iconPath) {
     return fs.readdirSync(iconPath).filter(file => /\.svg$/i.test(file));
 }
 
-function getMaxUnicode(iconPath) {
-    let max = 59904;
-    getSvgs(iconPath).forEach(function (svg) {
-        const exec = unicodeRE.exec(svg);
-        if (exec) {
-            const index = parseInt(exec[1], 16);
-            if (index > max) {
-                max = index;
-            }
-        }
-    });
-    return max;
-}
-
 function renameSvgs(iconPath) {
-    let begin = getMaxUnicode(iconPath) + 1;
     getSvgs(iconPath).forEach(function (svg) {
         if (unicodeRE.test(svg)) {
-            return;
+            const newPath = path.join(iconPath, svg.replace(unicodeRE, ''));
+            fs.renameSync(path.join(iconPath, svg), newPath);
+        } else {
+            if (svg.includes('-')) {
+                throw new Error(`svg fileName can't include -`)
+            }
         }
-        const newPath = iconPath + 'u' + begin.toString(16).toUpperCase() + '-' + svg;
-        fs.renameSync(iconPath + svg, newPath);
-        begin++;
     });
     return true;
 }
@@ -40,58 +27,62 @@ function renameSvgs(iconPath) {
 async function bundleIconFont(constPaths) {
     const {
         iconPath,
-        iconfontDistPath
+        iconfontDistPath,
+        uedTaskDir,
+        scssLocation
     } = constPaths;
 
     renameSvgs(iconPath)
 
     console.log('iconPath', iconPath)
     return new Promise((resolve, reject) => {
+
+        const fontCSSDest = path.join(uedTaskDir, 'src/css/common/iconfont.css');
+        const fontCodeReg = /\.icon-(.+?):before\s*{\s*content\s*:\s*"\\(\w+)/g;
+        const generateFileList = getSvgs(iconPath).map(fileName => path.join(iconPath, fileName));
         webfontsGenerator({
-            files: getSvgs(iconPath).map(fileName => path.join(iconPath, fileName)),
-            fontName: 'iconfont',
-            types: ['svg', 'ttf', 'eot', 'woff'],
+            files: generateFileList,
             dest: iconfontDistPath,
-            css: false,//Whether to generate CSS file.
-            normalize: true,
-            fontHeight: 1001,
-            startCodepoint: 0xEA01,
-            rename: (filePath) => {
-                const name = path.basename(filePath, path.extname(filePath))
-                const resName = name.split('-')[1];
-                return resName
-            }
+            css: true,//Whether to generate CSS file.
+            cssDest: fontCSSDest,
+            formatOptions: {
+                svg: {
+                    normalize: true,
+                    fontHeight: 1001
+                }
+            },
         }, function (error) {
             if (error) {
                 console.log('Fail!', error);
                 reject(error)
             } else {
-                resolve()
-                // console.log('Done!');
+                if (fs.existsSync(fontCSSDest)) {
+                    // 生成对应icons对象
+                    const ICON_DATA_STR = '$__iconfont__data';
+                    const ICON_FONT_FUNC = path.resolve(__dirname, '../lib/common-scss/_iconfont.scss');
+                    const icons = {};
+                    fs.readFileSync(fontCSSDest, {
+                        encoding: 'utf-8'
+                    }).replace(fontCodeReg, (match, $1, $2) => {
+                        icons[$1] = `\\${$2}`;
+                    });
+
+                    const content = [
+                        `${ICON_DATA_STR}: ` + JSON.stringify(icons, null, '\t').replace(/\{/g, '(').replace(/\}/g, ')').replace(/\\\\/g, '\\') + ';',
+                        fs.readFileSync(ICON_FONT_FUNC, {
+                            encoding: 'utf-8'
+                        }).toString().replace(/\r\n/ig, '\n')
+                    ].join('\n\n');
+                    console.log('scssLocation', scssLocation)
+
+                    fs.writeFileSync(path.join(scssLocation, '/_iconfont.scss'), content);
+                    console.log('iconfont generator done!');
+                    resolve();
+                }
+
             }
         })
     })
-}
-
-async function generateIconScss(constPaths) {
-    const { iconPath, scssLocation } = constPaths;
-
-    const pathname = path.resolve(__dirname, '../lib/common-scss/_iconfont.scss');
-
-    const icons = getSvgs(iconPath).reduce(function (iconfont, svg) {
-        const svgFileNameArr = path.basename(svg, path.extname(svg)).split('-')
-
-        return Object.defineProperty(iconfont, svgFileNameArr[1], {
-            value: svgFileNameArr[0].toLowerCase().replace('u', '\\'),
-            enumerable: true
-        })
-    }, {})
-
-    const content = [
-        '$__iconfont__data: ' + JSON.stringify(icons, null, '\t').replace(/\{/g, '(').replace(/\}/g, ')').replace(/\\\\/g, '\\') + ';',
-        fs.readFileSync(pathname).toString()
-    ].join('\n\n');
-    fs.writeFileSync(path.join(scssLocation, '/_iconfont.scss'), content);
 }
 
 module.exports = async (constPaths) => {
@@ -102,7 +93,6 @@ module.exports = async (constPaths) => {
         return Promise.resolve();
     }
     await bundleIconFont(constPaths);
-    await generateIconScss(constPaths);
 
     let timer = null;
     let initWatch = true;
